Backend/requirements.txt
amqp               5.1.1    
async-timeout      4.0.2    
backports.zoneinfo 0.2.1    
billiard           4.1.0    
blinker            1.6.2    
cachelib           0.9.0    
celery             5.3.1    
certifi            2023.7.22
charset-normalizer 3.2.0    
click              8.1.6    
click-didyoumean   0.3.0    
click-plugins      1.1.1    
click-repl         0.3.0    
Flask              2.3.2    
Flask-Caching      2.0.2    
Flask-Cors         4.0.0    
Flask-JWT-Extended 4.5.2    
Flask-Mail         0.9.1    
flask-sqlalchemy   3.0.5    
greenlet           2.0.2    
idna               3.4      
importlib-metadata 6.8.0    
itsdangerous       2.1.2    
Jinja2             3.1.2    
kombu              5.3.1    
MarkupSafe         2.1.3    
numpy              1.24.4   
pandas             2.0.3    
pip                20.0.2   
pkg-resources      0.0.0    
prompt-toolkit     3.0.39   
PyJWT              2.8.0    
python-dateutil    2.8.2    
python-dotenv      1.0.0    
pytz               2023.3   
redis              4.6.0    
requests           2.31.0   
setuptools         44.0.0   
six                1.16.0   
SQLAlchemy         2.0.19   
typing-extensions  4.7.1    
tzdata             2023.3   
urllib3            2.0.4    
vine               5.0.0    
wcwidth            0.2.6    
Werkzeug           2.3.6    
wheel              0.34.2   
zipp               3.16.2   
